requirements.txt
langchain
huggingface_hub
python-dotenv
streamlit